restructure components, move to separate files
