+++ catalog/app/containers/Search/Search.tsx
@@ -1,8 +1,6 @@
 import cx from 'classnames'
-import jsonpath from 'jsonpath'
 import invariant from 'invariant'
 import * as React from 'react'
-import * as RR from 'react-router-dom'
 import { useDebouncedCallback } from 'use-debounce'
 import * as M from '@material-ui/core'
 import * as Lab from '@material-ui/lab'
@@ -13,19 +11,18 @@ import Skeleton from 'components/Skeleton'
 import * as GQL from 'utils/GraphQL'
 import * as JSONPointer from 'utils/JSONPointer'
 import MetaTitle from 'utils/MetaTitle'
-import * as NamedRoutes from 'utils/NamedRoutes'
 import assertNever from 'utils/assertNever'
 import * as Format from 'utils/format'
-import { readableBytes } from 'utils/string'
-import type { Json, JsonRecord } from 'utils/types'
 
 import * as SearchUIModel from './model'
 import AssistantContext from './AssistantContext'
 import BucketSelector from './Buckets'
-import * as Hit from './Hit'
 import ResultTypeSelector from './ResultType'
 import { EmptyResults, ResultsSkeleton, SearchError } from './Results'
 import SortSelector from './Sort'
+import ListView, { ListViewProps } from './Views/List'
+import TableView, { TableViewProps } from './Views/Table'
+import { packageFilterLabels } from './i18n'
 
 function useMobileView() {
   const t = M.useTheme()
@@ -369,16 +366,6 @@ const WIDGETS = {
 function FilterWidget(props: FilterWidgetProps) {
   const Widget = WIDGETS[props.state._tag]
   return <Widget {...(props as $TSFixMe)} />
-}
-
-const packageFilterLabels = {
-  modified: 'Last modified',
-  size: 'Cumulative package size',
-  name: 'Package name',
-  hash: 'Package hash',
-  entries: 'Total number of files in the package',
-  comment: 'Package revision comment',
-  workflow: 'Workflow',
 }
 
 interface PackagesFilterActivatorProps {
@@ -1066,45 +1053,6 @@ function Filters({ className }: FiltersProps) {
   )
 }
 
-interface SearchHitProps {
-  hit: SearchUIModel.SearchHit
-  showBucket: boolean
-  showRevision: boolean
-}
-
-function SearchHit({ hit, showBucket, showRevision }: SearchHitProps) {
-  switch (hit.__typename) {
-    case 'SearchHitObject':
-      return (
-        <Hit.Object
-          showBucket={showBucket}
-          hit={hit}
-          data-testid="search-hit"
-          data-search-hit-type="file"
-          data-search-hit-bucket={hit.bucket}
-          data-search-hit-path={hit.key}
-        />
-      )
-
-    case 'SearchHitPackage':
-      return (
-        <Hit.Package
-          showBucket={showBucket}
-          showRevision={showRevision}
-          hit={hit}
-          data-testid="search-hit"
-          data-search-hit-type="package"
-          data-search-hit-bucket={hit.bucket}
-          data-search-hit-package-name={hit.name}
-          data-search-hit-package-hash={hit.hash}
-        />
-      )
-
-    default:
-      assertNever(hit)
-  }
-}
-
 const useLoadNextPageStyles = M.makeStyles((t) => ({
   root: {
     padding: t.spacing(1, 0),
@@ -1138,227 +1086,6 @@ function LoadNextPage({ className, loading = false, onClick }: LoadNextPageProps
         Load more
       </M.Button>
     </div>
-  )
-}
-
-interface ViewProps {
-  hits: readonly SearchUIModel.SearchHit[]
-}
-
-interface ListViewProps extends ViewProps {
-  showBucket: boolean
-  showRevision: boolean
-}
-
-function ListView({ hits, showBucket, showRevision }: ListViewProps) {
-  return (
-    <>
-      {hits.map((hit) => (
-        <SearchHit
-          key={hit.id}
-          hit={hit}
-          showBucket={showBucket}
-          showRevision={showRevision}
-        />
-      ))}
-    </>
-  )
-}
-
-const columnLabels = {
-  modified: 'Modified',
-  size: 'Size',
-  name: 'Name',
-  hash: 'Hash',
-  entries: 'Entries number',
-  comment: 'Comment',
-  workflow: 'Workflow',
-}
-
-const NoValue = () => (
-  <M.Box display="inline-block" width={16}>
-    <M.Divider />
-  </M.Box>
-)
-
-const isJsonRecord = (obj: Json): obj is JsonRecord =>
-  obj != null && typeof obj === 'object' && !Array.isArray(obj)
-
-interface TableViewSystemMetaProps {
-  hit: SearchUIModel.SearchHitPackage
-  filter: SearchUIModel.FilterStateForResultType<SearchUIModel.ResultType.QuiltPackage>['order'][number]
-}
-
-function TableViewSystemMeta({ hit, filter }: TableViewSystemMetaProps) {
-  switch (filter) {
-    case 'workflow':
-      return hit.workflow ? hit.workflow.id : <NoValue />
-    case 'size':
-      return readableBytes(hit.size)
-    case 'modified':
-      return <Format.Relative value={hit.modified} />
-    default:
-      return null
-  }
-}
-
-interface TableViewUserMetaProps {
-  meta: JsonRecord
-  pointer: JSONPointer.Pointer
-}
-
-function TableViewUserMeta({ meta, pointer }: TableViewUserMetaProps) {
-  if (!isJsonRecord(meta))
-    return (
-      <M.Tooltip title={`${meta}`}>
-        <M.Icon color="disabled" fontSize="small" style={{ verticalAlign: 'middle' }}>
-          error_outline
-        </M.Icon>
-      </M.Tooltip>
-    )
-  const value = jsonpath.value(meta, JSONPointer.toJsonPath(pointer))
-  return value || <NoValue />
-}
-
-const useTableViewHitStyles = M.makeStyles((t) => ({
-  cell: {
-    whiteSpace: 'nowrap',
-  },
-  head: {
-    display: 'flex',
-    alignItems: 'center',
-    position: 'relative',
-    '&::after': {
-      content: '""',
-      position: 'absolute',
-      right: t.spacing(-3),
-      top: t.spacing(1),
-      bottom: t.spacing(1),
-      background: t.palette.divider,
-      width: '1px',
-    },
-    '&:hover $headButton': {
-      opacity: 1,
-    },
-  },
-  headButton: {
-    marginLeft: t.spacing(2),
-    opacity: 0,
-    transition: 'opacity ease 0.3s',
-  },
-}))
-
-interface TableViewPackageProps {
-  hit: SearchUIModel.SearchHitPackage
-}
-
-function TableViewPackage({ hit }: TableViewPackageProps) {
-  const { state } = SearchUIModel.use(SearchUIModel.ResultType.QuiltPackage)
-  const meta = hit.meta ? JSON.parse(hit.meta) : {}
-  const classes = useTableViewHitStyles()
-  const { urls } = NamedRoutes.use()
-  return (
-    <M.TableRow hover>
-      <M.TableCell className={classes.cell}>
-        <RR.Link to={urls.bucketPackageTree(hit.bucket, hit.name, hit.hash)}>
-          {hit.name}
-        </RR.Link>
-      </M.TableCell>
-      {state.filter.order.map((filter) => (
-        <M.TableCell
-          className={classes.cell}
-          data-search-hit-filter={filter}
-          key={filter}
-        >
-          <TableViewSystemMeta hit={hit} filter={filter} />
-        </M.TableCell>
-      ))}
-      {Array.from(state.userMetaFilters.filters.keys()).map((key) => (
-        <M.TableCell className={classes.cell} data-search-hit-meta={key} key={key}>
-          <TableViewUserMeta meta={meta} pointer={key} />
-        </M.TableCell>
-      ))}
-    </M.TableRow>
-  )
-}
-
-interface TableViewObjectProps {
-  hit: SearchUIModel.SearchHitObject
-}
-
-function TableViewObject({ hit }: TableViewObjectProps) {
-  return (
-    <M.TableRow hover>
-      <M.TableCell>{hit.key}</M.TableCell>
-    </M.TableRow>
-  )
-}
-
-interface TableViewHitProps {
-  hit: SearchUIModel.SearchHit
-}
-
-function TableViewHit({ hit }: TableViewHitProps) {
-  switch (hit.__typename) {
-    case 'SearchHitObject':
-      return <TableViewObject hit={hit} />
-    case 'SearchHitPackage':
-      return <TableViewPackage hit={hit} />
-    default:
-      assertNever(hit)
-  }
-}
-
-interface TableViewProps extends ViewProps {}
-
-function TableView({ hits }: TableViewProps) {
-  const { actions, state } = SearchUIModel.use(SearchUIModel.ResultType.QuiltPackage)
-  const classes = useTableViewHitStyles()
-  return (
-    <M.Table size="small">
-      <M.TableHead>
-        <M.TableRow>
-          <M.TableCell className={classes.cell}>
-            <div className={classes.head}>Name</div>
-          </M.TableCell>
-          {state.filter.order.map((filter) => (
-            <M.TableCell key={filter} className={classes.cell}>
-              <div className={classes.head}>
-                <M.Tooltip title={packageFilterLabels[filter]}>
-                  <span>{columnLabels[filter]}</span>
-                </M.Tooltip>
-                <M.IconButton
-                  className={classes.headButton}
-                  size="small"
-                  onClick={() => actions.deactivatePackagesFilter(filter)}
-                >
-                  <M.Icon fontSize="inherit">close</M.Icon>
-                </M.IconButton>
-              </div>
-            </M.TableCell>
-          ))}
-          {Array.from(state.userMetaFilters.filters.keys()).map((key) => (
-            <M.TableCell key={key} className={classes.cell}>
-              <div className={classes.head}>
-                {key}
-                <M.IconButton
-                  className={classes.headButton}
-                  size="small"
-                  onClick={() => actions.deactivatePackagesMetaFilter(key)}
-                >
-                  <M.Icon fontSize="inherit">close</M.Icon>
-                </M.IconButton>
-              </div>
-            </M.TableCell>
-          ))}
-        </M.TableRow>
-      </M.TableHead>
-      <M.TableBody>
-        {hits.map((hit) => (
-          <TableViewHit key={hit.id} hit={hit} />
-        ))}
-      </M.TableBody>
-    </M.Table>
   )
 }
