+++ catalog/app/containers/Search/i18n.ts
@@ -1,0 +1,21 @@
+// Descriptive filter labels
+export const packageFilterLabels = {
+  modified: 'Last modified',
+  size: 'Cumulative package size',
+  name: 'Package name',
+  hash: 'Package hash',
+  entries: 'Total number of files in the package',
+  comment: 'Package revision comment',
+  workflow: 'Workflow',
+}
+
+// Shortened filter labels
+export const columnLabels = {
+  modified: 'Modified',
+  size: 'Size',
+  name: 'Name',
+  hash: 'Hash',
+  entries: 'Entries number',
+  comment: 'Comment',
+  workflow: 'Workflow',
+}
